# flask/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Access API",
        "description": "API endpoints for UGV-Based Disaster Management System",
        "version": "1.0"
    },
    "basePath": "/",
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Bearer token"
        }
    },
    "paths": {
...
        "/api/devices": {
            "post": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Registration [ADMIN - ROBOT]",
                "description": "Register a new device\nName and MAC Address must be unique.\nThis endpoint is called from the device itself, not the mobile app.",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Device data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "Device1",
                                    "required": true
                                },
                                "password": {
                                    "type": "string",
                                    "example": "Password@123",
                                    "required": true
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16",
                                    "required": true
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                },
                                "broker_id": {
                                    "type": "string",
                                    "example": "6055a0ae20e90e08641e22ef",
                                    "required": false
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Device registered successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device registered successfully."
                                },
                                "device_id": {
                                    "type": "string",
                                    "example": "6055a0ae20e90e08641e22ef"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "409": {
                        "description": "Conflict - A device with the same MAC address is already registered"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/rtmp_auth": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Authentication Proxy for RTMP Server",
                "description": "Login with name and password for RTMP Server",
                "parameters": [
                    {
                        "in": "query",
                        "name": "name",
                        "type": "string",
                        "required": true,
                        "default": "user123"
                    },
                    {
                        "in": "query",
                        "name": "password",
                        "type": "string",
                        "required": true,
                        "default": "New_password123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device logged in successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid email or password"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/devices/{device_id}": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Info Retrievement [ADMIN]",
                "description": "Get Device Information for Admin interface",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to retrieve",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON object of device information",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "device_id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Robot1"
                                },
                                "mac": {
                                    "type": "string",
                                    "example": "00:0a:95:9d:68:16"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                },
                                "status": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid token"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Info Update [ADMIN]",
                "description": "update name, or password for the Device\n\nDevice Types:\n1: UGV\n2: UAV\n3: DOG\n4: CHARGING_STATION\n 5: BROKER",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to update",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Device data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "type": "string",
                                    "example": "new_name"
                                },
                                "old_password": {
                                    "type": "string",
                                    "example": "Old_password123"
                                },
                                "new_password": {
                                    "type": "string",
                                    "example": "New_password123"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device information is updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device information is updated successfully."
                                },
                                "name": {
                                    "type": "string",
                                    "example": "Robot1"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Deactivation [ADMIN]",
                "description": "Deactivate a device",
                "parameters": [
                    {
                        "name": "device_id",
                        "in": "path",
                        "description": "ID of the device to deactivate",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device is deactivated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Device is deactivated successfully."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized Error - Admin access is required"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - Device not found"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/all": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device List Retrievement [ADMIN]",
                "description": "Get list of devices with pagenation for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page-number",
                        "type": "integer",
                        "required": false,
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "in": "query",
                        "name": "page-size",
                        "type": "integer",
                        "required": false,
                        "description": "Number of items per page",
                        "default": 5
                    },
                    {
                        "in": "query",
                        "name": "name",
                        "type": "string",
                        "required": false,
                        "description": "filter for device name",
                        "default": "device"
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "type": "array",
                        "required": false,
                        "description": "filter for device status",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "type": "array",
                        "required": false,
                        "description": "filter for device type",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "mission",
                        "type": "string",
                        "required": false,
                        "description": "[mission_id] devices for mission to include in the list"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "6055a0ae80e90e08641e22ef"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Robot1"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/count": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Device Count [ADMIN]",
                "description": "Get the count of devices with filters for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "description": "filter for the user status",
                        "type": "array",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "type": "array",
                        "required": false,
                        "description": "filter for the user type",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "array",
                                    "example": []
                                },
                                "type": {
                                    "type": "array",
                                    "example": [
                                        1,
                                        2
                                    ]
                                },
                                "count": {
                                    "type": "int",
                                    "example": 10
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/devices/broker_id": {
            "get": {
                "tags": [
                    "Device Management"
                ],
                "summary": "Broker ID [ADMIN - ROBOT]",
                "description": "Get the broker id by using its mac address",
                "parameters": [
                    {
                        "in": "query",
                        "name": "mac",
                        "required": true,
                        "description": "mac of the broker",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "broker_id": {
                                    "type": "string",
                                    "example": "6055a0ae20e90e08641e22ef"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    }
}